Uncomment the iOS build variants and get everything building successfully again.
The build was in a dodgy state due to the age of some of the libraries
being used (versions from back when KMM was very new). I've updated to
a much more recent version of multiplatform itself and ktor to go with
it.

Some changes to imports and some minor code changes were required. There
are now some warnings about deprecated things being used from the IO utils
package but I will fix them separately.